Organize Common into Patching and Models folders
Move the patch-file engine and its internals (Ksp2Patch, FileInformation,
PatchManifest, PatchOperation) into Patching/, and the release-feed schema
(ReleaseManifest, ReleasePatch, PatchRequirement) into Models/, with
namespaces to match. Rename the Common.Service namespace to Services,
update all consumers, and drop a dead using in CacheService.
