+++ Ksp2Redux.Tools.Installer/App.xaml
@@ -1,7 +1,6 @@
 ﻿<Application x:Class="Ksp2Redux.Tools.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:Ksp2Redux.Tools"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
         <ResourceDictionary>
