+++ Ksp2Redux.Tools.Installer/AssemblyInfo.cs
@@ -1,10 +1,10 @@
 using System.Windows;
 
 [assembly: ThemeInfo(
-    ResourceDictionaryLocation.None, //where theme specific resource dictionaries are located
+    ResourceDictionaryLocation.None, //where theme-specific resource dictionaries are located
     //(used if a resource is not found in the page,
     // or application resource dictionaries)
     ResourceDictionaryLocation.SourceAssembly //where the generic resource dictionary is located
     //(used if a resource is not found in the page,
-    // app, or any theme specific resource dictionaries)
+    // app, or any theme-specific resource dictionaries)
 )]
