+++ Ksp2Redux.Tools.Installer/MainWindow.xaml
@@ -3,7 +3,6 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Ksp2Redux.Tools"
         mc:Ignorable="d"
         Title="KSP2 Redux Installer" Height="450" Width="800">
         <Grid>
