+++ Ksp2Redux.Tools.Launcher.Test.MockGame/Program.cs
@@ -1,5 +1,5 @@
 ﻿
 // Mock application used in Ksp2Redux.Tools.Launcher.Test to act as the game
-// Doesn't contain a VersionID, this one is mocked in the test directly
+// Doesn't contain a VersionID; this one is mocked in the test directly
 
-Console.WriteLine($"Mock application for KSP2 Redux");
+Console.WriteLine("Mock application for KSP2 Redux");
