+++ Ksp2Redux.Tools.Launcher.Test/HeadlessTests/DownloadTest.cs
@@ -10,7 +10,8 @@ using Avalonia.Input;
 using Avalonia.Threading;
 using Avalonia.VisualTree;
 using CodeHollow.FeedReader;
-using Ksp2Redux.Tools.Common;
+using Ksp2Redux.Tools.Common.Models;
+using Ksp2Redux.Tools.Common.Wrappers;
 using Ksp2Redux.Tools.Launcher.Controls;
 using Ksp2Redux.Tools.Launcher.Models;
 using Ksp2Redux.Tools.Launcher.ViewModels;
