+++ Ksp2Redux.Tools.Launcher.Test/Ksp2PatchPathTraversalTest.cs
@@ -1,8 +1,8 @@
 using System.IO.Abstractions;
 using System.IO.Compression;
 using System.Text;
-using Ksp2Redux.Tools.Common;
-using Ksp2Redux.Tools.Common.Service;
+using Ksp2Redux.Tools.Common.Patching;
+using Ksp2Redux.Tools.Common.Wrappers;
 using Moq;
 using Testably.Abstractions.Testing;
