+++ Ksp2Redux.Tools.Launcher.Test/ManifestReleasesFeedDownloadPatchTest.cs
@@ -1,6 +1,6 @@
 using System.Net;
 using System.Security.Cryptography;
-using Ksp2Redux.Tools.Common;
+using Ksp2Redux.Tools.Common.Models;
 using Ksp2Redux.Tools.Launcher.Models;
 using Ksp2Redux.Tools.Launcher.Services;
 using Moq;
