+++ Ksp2Redux.Tools.Launcher.Test/ManifestReleasesFeedUpdateManifestTest.cs
@@ -1,4 +1,4 @@
-using Ksp2Redux.Tools.Common;
+using Ksp2Redux.Tools.Common.Models;
 using Ksp2Redux.Tools.Launcher.Models;
 using Ksp2Redux.Tools.Launcher.Services;
 using Moq;
