+++ Ksp2Redux.Tools.Launcher/Models/ReleasePatchExtensions.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Linq;
-using Ksp2Redux.Tools.Common;
+using Ksp2Redux.Tools.Common.Models;
 
 namespace Ksp2Redux.Tools.Launcher.Models;
